Refactor coordinate parsing, add tests
Refactor the script to make it more Python and eliminate some
of the nesting.

Also factor out the logic to parse the coordinates into a
separate function and add unit tests to make it easier to
implement a fix to the broken parsing in the future.

Use named groups in the matching regexes, and have separate
regexes for the full and shortened Google Maps URLs because
it just makes the code more readable.

Addresses #4