Handle shortened URLs that don't redirect all the way
It seems that many shortened URLs are getting resolved to
https://www.google.com/sorry/index, which prompts for the user
to complete a CAPTCHA before redirecting to the final Google Maps
URL. The final URL, along with the coordinates is available
through the `continue` query parameter, but the final URL
is encoded. So, the regex that checks for the coordinates
needs to match both '@' and the encoding for that character,
'%40'.

Fixes #4