Fix failure to resolve shortened URLs (#5)
* Update script to make further development easier

Update the script and documentation to make it easier to do further
development and avoid crashing out.

- Make this a proper Python package that installs a command-line script.
- Exit if configuration environment variables aren't set.
- Using Python's built-in logging to write error messages to standard
  error instead of standard output.
- Document assumptions of the app, the configuration environment
  variables, where to find docs to get required Trello credentials, and
  steps to install the package for local development.

There are a few breaking changes:

- Since this is now a Python package that installs a command-line script
  to the user's path, the script is now run with `trello-to-geojson`
  instead of `python get_board.py`.
- The program no longer hard codes the output file and instead writes
  the GeoJSON to standard output. To save to a file, redirect the output
  like `trello-to-geojson > out.geojson`.

Addresses #4, #3

* Refactor coordinate parsing, add tests

Refactor the script to make it more Python and eliminate some
of the nesting.

Also factor out the logic to parse the coordinates into a
separate function and add unit tests to make it easier to
implement a fix to the broken parsing in the future.

Use named groups in the matching regexes, and have separate
regexes for the full and shortened Google Maps URLs because
it just makes the code more readable.

Addresses #4

* Handle shortened URLs that don't redirect all the way

It seems that many shortened URLs are getting resolved to
https://www.google.com/sorry/index, which prompts for the user
to complete a CAPTCHA before redirecting to the final Google Maps
URL. The final URL, along with the coordinates is available
through the `continue` query parameter, but the final URL
is encoded. So, the regex that checks for the coordinates
needs to match both '@' and the encoding for that character,
'%40'.

Fixes #4

* Make the Slack notifier a proper command-line script

- Define it in pyproject.toml
- Parameterize the channel ID and message

This breaks the way the script is run. See the README for new
usage syntax.

Addresses #3, #4

* Linting